requirements.txt
# KriptoBot - Multi-User Futures Trading Platform
# Python 3.10+

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Authentication
PyJWT==2.8.0

# HTTP İstekleri
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Telegram
telethon==1.34.0
python-telegram-bot==20.7

# Google Gemini AI
google-generativeai==0.3.2

# Veritabanı
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# Teknik Analiz
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
pandas-ta==0.3.14b

# Zamanlayıcı
apscheduler==3.10.4
schedule==1.2.1

# Async
asyncio-throttle==1.0.2

# Ortam Değişkenleri
python-dotenv==1.0.0

# Loglama
loguru==0.7.2

# Yardımcı
python-dateutil==2.8.2
pytz==2023.3


